fix: tle 无法检测

+++ src/autohack/core/util.py
@@ -1,5 +1,5 @@
 from autohack.core.path import *
-import subprocess, threading, psutil, os
+import subprocess, threading, psutil, time, os
 
 
 class CodeRunner:
@@ -19,18 +19,30 @@ class CodeRunner:
             self.stderr = stderr
 
     def __init__(self):
+        self.timeOut = False
         self.memoryOut = False
 
-    def memoryMonitor(self, pid: int, memoryLimit: int) -> None:
+    def memoryMonitor(
+        self, pid: int, timeLimit: float | None, memoryLimit: int | None
+    ) -> None:
         try:
             psutilProcess = psutil.Process(pid)
         except psutil.NoSuchProcess:
             # 跑的太他妈快了，没测到
             return
+        startTime = psutilProcess.create_time()
         while True:
             try:
+                # psutilProcess.cpu_times();
+                if timeLimit is not None and time.time() - startTime > timeLimit:
+                    self.timeOut = True
+                    psutilProcess.kill()
+                    return
                 # 测出来是资源监视器内存中提交那栏 *1024
-                if psutilProcess.memory_info().vms > memoryLimit:
+                if (
+                    memoryLimit is not None
+                    and psutilProcess.memory_info().vms > memoryLimit
+                ):
                     self.memoryOut = True
                     psutilProcess.kill()
                     return
@@ -41,35 +53,35 @@ class CodeRunner:
         self,
         *popenargs,
         inputContent: bytes | None = None,
-        timeLimit: int | None = None,
+        timeLimit: float | None = None,
         memoryLimit: int | None = None,
         **kwargs,
     ) -> Result:
-        timeOut = False
         returnCode = 0
         stdout = None
         stderr = None
         with subprocess.Popen(*popenargs, **kwargs) as process:
-            if memoryLimit is not None:
-                monitor = threading.Thread(
-                    target=self.memoryMonitor,
-                    args=(
-                        process.pid,
-                        memoryLimit,
-                    ),
-                )
-                monitor.start()
-            try:
-                stdout, stderr = process.communicate(inputContent, timeout=timeLimit)
-            except subprocess.TimeoutExpired:
-                process.kill()
-                if mswindows():
-                    stdout, stderr = process.communicate()
-                else:
-                    process.wait()
-                timeOut = True
+            monitor = threading.Thread(
+                target=self.memoryMonitor,
+                args=(
+                    process.pid,
+                    timeLimit,
+                    memoryLimit,
+                ),
+            )
+            monitor.start()
+            stdout, stderr = process.communicate(inputContent)
+            # try:
+            #     stdout, stderr = process.communicate(inputContent, timeout=timeLimit)
+            # except subprocess.TimeoutExpired:
+            #     process.kill()
+            #     if mswindows():
+            #         stdout, stderr = process.communicate()
+            #     else:
+            #         process.wait()
+            #     timeOut = True
             returnCode = process.poll()
-        return self.Result(timeOut, self.memoryOut, returnCode, stdout, stderr)
+        return self.Result(self.timeOut, self.memoryOut, returnCode, stdout, stderr)
 
 
 def checkDirectoryExists(directory: str) -> None:
